# Bump osmium to version 2.20.0