… third_party/libosmium/.clang-tidy
… third_party/libosmium/.github/actions/cmake-windows/action.yml
… third_party/libosmium/.github/actions/install-ubuntu/action.yml
… third_party/libosmium/.github/workflows/ci.yml
… third_party/libosmium/.github/workflows/clang-tidy.yml
… third_party/libosmium/.ycm_extra_conf.py
… third_party/libosmium/CHANGELOG.md
… third_party/libosmium/CMakeLists.txt
… third_party/libosmium/NOTES_FOR_DEVELOPERS.md
… third_party/libosmium/README.md
… third_party/libosmium/benchmarks/osmium_benchmark_static_vs_dynamic_index.cpp
… third_party/libosmium/benchmarks/osmium_benchmark_write_pbf.cpp
… third_party/libosmium/build-appveyor.bat
… third_party/libosmium/build-msys2.bat
… third_party/libosmium/examples/osmium_amenity_list.cpp
… third_party/libosmium/examples/osmium_area_test.cpp
… third_party/libosmium/examples/osmium_change_tags.cpp
… third_party/libosmium/examples/osmium_count.cpp
… third_party/libosmium/examples/osmium_create_pois.cpp
… third_party/libosmium/examples/osmium_debug.cpp
… third_party/libosmium/examples/osmium_dump_internal.cpp
… third_party/libosmium/examples/osmium_filter_discussions.cpp
… third_party/libosmium/examples/osmium_index_lookup.cpp
… third_party/libosmium/examples/osmium_read.cpp
… third_party/libosmium/examples/osmium_read_with_progress.cpp
… third_party/libosmium/examples/osmium_tiles.cpp
… third_party/libosmium/include/osmium/area/assembler.hpp
… third_party/libosmium/include/osmium/area/assembler_config.hpp
… third_party/libosmium/include/osmium/area/assembler_legacy.hpp
… third_party/libosmium/include/osmium/area/detail/basic_assembler.hpp