<!-- BENCHMARK_RESULTS_START -->
<details><summary><h2>Benchmark Results</h2></summary>

| Benchmark | Base | PR |
|-----------|------|----|
| alias | aliased u32: 1155.5<br/>plain u32: 1143.43<br/>aliased double: 1259.59<br/>plain double: 1278 | aliased u32: 1210.63<br/>plain u32: 1176.24<br/>aliased double: 1264.12<br/>plain double: 1244.09 |
| e2e_match_ch | Ops: 41.22 ± 0.07 ops/s. Best: 41.32 ops/s<br/>Total: 3177.97ms ± 5.85ms. Best: 3170.42ms<br/>Min time: 2.17ms ± 0.03ms<br/>Mean time: 24.26ms ± 0.05ms<br/>Median time: 17.85ms ± 0.11ms<br/>95th percentile: 83.02ms ± 0.50ms<br/>99th percentile: 102.68ms ± 0.49ms<br/>Max time: 107.27ms ± 0.62ms | Ops: 41.69 ± 0.15 ops/s. Best: 41.97 ops/s<br/>Total: 3142.47ms ± 12.47ms. Best: 3121.02ms<br/>Min time: 2.18ms ± 0.05ms<br/>Mean time: 23.99ms ± 0.10ms<br/>Median time: 18.22ms ± 0.09ms<br/>95th percentile: 80.83ms ± 0.34ms<br/>99th percentile: 99.30ms ± 0.56ms<br/>Max time: 106.69ms ± 0.74ms |
| e2e_match_mld | Ops: 64.06 ± 0.24 ops/s. Best: 64.31 ops/s<br/>Total: 2045.34ms ± 7.55ms. Best: 2037.03ms<br/>Min time: 1.72ms ± 0.02ms<br/>Mean time: 15.61ms ± 0.06ms<br/>Median time: 8.09ms ± 0.08ms<br/>95th percentile: 52.27ms ± 0.13ms<br/>99th percentile: 60.64ms ± 0.14ms<br/>Max time: 69.83ms ± 0.42ms | Ops: 64.04 ± 0.25 ops/s. Best: 64.38 ops/s<br/>Total: 2045.77ms ± 7.72ms. Best: 2034.79ms<br/>Min time: 1.73ms ± 0.01ms<br/>Mean time: 15.61ms ± 0.06ms<br/>Median time: 8.09ms ± 0.07ms<br/>95th percentile: 52.12ms ± 0.34ms<br/>99th percentile: 61.73ms ± 1.77ms<br/>Max time: 70.62ms ± 1.89ms |
| e2e_nearest_ch | Ops: 867.49 ± 2.74 ops/s. Best: 872.64 ops/s<br/>Total: 1152.76ms ± 4.19ms. Best: 1145.95ms<br/>Min time: 0.98ms ± 0.01ms<br/>Mean time: 1.15ms ± 0.00ms<br/>Median time: 1.06ms ± 0.00ms<br/>95th percentile: 1.61ms ± 0.00ms<br/>99th percentile: 1.65ms ± 0.01ms<br/>Max time: 5.96ms ± 2.79ms | Ops: 875.54 ± 3.33 ops/s. Best: 881.73 ops/s<br/>Total: 1142.19ms ± 4.77ms. Best: 1134.14ms<br/>Min time: 0.98ms ± 0.00ms<br/>Mean time: 1.14ms ± 0.00ms<br/>Median time: 1.06ms ± 0.00ms<br/>95th percentile: 1.59ms ± 0.01ms<br/>99th percentile: 1.63ms ± 0.01ms<br/>Max time: 5.83ms ± 2.83ms |
| e2e_nearest_mld | Ops: 860.05 ± 8.31 ops/s. Best: 875.29 ops/s<br/>Total: 1163.12ms ± 12.06ms. Best: 1142.48ms<br/>Min time: 0.98ms ± 0.00ms<br/>Mean time: 1.16ms ± 0.01ms<br/>Median time: 1.07ms ± 0.01ms<br/>95th percentile: 1.61ms ± 0.02ms<br/>99th percentile: 1.75ms ± 0.14ms<br/>Max time: 6.87ms ± 2.98ms | Ops: 862.15 ± 6.79 ops/s. Best: 872.95 ops/s<br/>Total: 1159.73ms ± 8.84ms. Best: 1145.54ms<br/>Min time: 0.99ms ± 0.00ms<br/>Mean time: 1.16ms ± 0.01ms<br/>Median time: 1.07ms ± 0.01ms<br/>95th percentile: 1.61ms ± 0.01ms<br/>99th percentile: 1.66ms ± 0.01ms<br/>Max time: 6.26ms ± 3.08ms |
| e2e_route_ch | Ops: 364.66 ± 2.76 ops/s. Best: 371.30 ops/s<br/>Total: 2742.61ms ± 20.29ms. Best: 2693.22ms<br/>Min time: 1.19ms ± 0.01ms<br/>Mean time: 2.74ms ± 0.02ms<br/>Median time: 2.76ms ± 0.02ms<br/>95th percentile: 3.64ms ± 0.03ms<br/>99th percentile: 4.04ms ± 0.05ms<br/>Max time: 7.07ms ± 2.68ms | Ops: 379.00 ± 0.82 ops/s. Best: 380.62 ops/s<br/>Total: 2638.57ms ± 5.68ms. Best: 2627.30ms<br/>Min time: 1.18ms ± 0.01ms<br/>Mean time: 2.64ms ± 0.01ms<br/>Median time: 2.66ms ± 0.01ms<br/>95th percentile: 3.48ms ± 0.01ms<br/>99th percentile: 3.87ms ± 0.03ms<br/>Max time: 6.59ms ± 2.49ms |
| e2e_route_mld | Ops: 287.05 ± 6.56 ops/s. Best: 300.48 ops/s<br/>Total: 3487.42ms ± 85.57ms. Best: 3328.06ms<br/>Min time: 1.21ms ± 0.01ms<br/>Mean time: 3.49ms ± 0.09ms<br/>Median time: 3.53ms ± 0.08ms<br/>95th percentile: 4.83ms ± 0.12ms<br/>99th percentile: 5.32ms ± 0.14ms<br/>Max time: 8.45ms ± 2.76ms | Ops: 313.82 ± 3.68 ops/s. Best: 318.72 ops/s<br/>Total: 3188.59ms ± 37.21ms. Best: 3137.54ms<br/>Min time: 1.17ms ± 0.01ms<br/>Mean time: 3.19ms ± 0.04ms<br/>Median time: 3.23ms ± 0.04ms<br/>95th percentile: 4.34ms ± 0.06ms<br/>99th percentile: 4.87ms ± 0.12ms<br/>Max time: 8.01ms ± 1.67ms |
| e2e_table_ch | Ops: 318.94 ± 0.62 ops/s. Best: 319.85 ops/s<br/>Total: 3135.45ms ± 6.40ms. Best: 3126.45ms<br/>Min time: 1.55ms ± 0.04ms<br/>Mean time: 3.14ms ± 0.01ms<br/>Median time: 3.13ms ± 0.02ms<br/>95th percentile: 4.37ms ± 0.02ms<br/>99th percentile: 4.71ms ± 0.05ms<br/>Max time: 8.38ms ± 2.54ms | Ops: 316.46 ± 0.57 ops/s. Best: 317.33 ops/s<br/>Total: 3159.86ms ± 5.83ms. Best: 3151.30ms<br/>Min time: 1.60ms ± 0.04ms<br/>Mean time: 3.16ms ± 0.01ms<br/>Median time: 3.15ms ± 0.01ms<br/>95th percentile: 4.39ms ± 0.02ms<br/>99th percentile: 4.73ms ± 0.07ms<br/>Max time: 8.54ms ± 2.52ms |
| e2e_table_mld | Ops: 110.64 ± 0.67 ops/s. Best: 111.20 ops/s<br/>Total: 9040.14ms ± 57.46ms. Best: 8992.65ms<br/>Min time: 3.66ms ± 0.08ms<br/>Mean time: 9.04ms ± 0.06ms<br/>Median time: 9.00ms ± 0.05ms<br/>95th percentile: 13.81ms ± 0.07ms<br/>99th percentile: 14.56ms ± 0.16ms<br/>Max time: 18.68ms ± 3.31ms | Ops: 112.14 ± 0.10 ops/s. Best: 112.31 ops/s<br/>Total: 8917.48ms ± 8.02ms. Best: 8903.94ms<br/>Min time: 3.60ms ± 0.07ms<br/>Mean time: 8.92ms ± 0.01ms<br/>Median time: 8.88ms ± 0.04ms<br/>95th percentile: 13.68ms ± 0.03ms<br/>99th percentile: 14.33ms ± 0.05ms<br/>Max time: 16.63ms ± 2.14ms |
| e2e_trip_ch | Ops: 97.30 ± 0.38 ops/s. Best: 97.89 ops/s<br/>Total: 10276.41ms ± 39.30ms. Best: 10215.96ms<br/>Min time: 1.52ms ± 0.09ms<br/>Mean time: 10.28ms ± 0.04ms<br/>Median time: 9.72ms ± 0.03ms<br/>95th percentile: 18.48ms ± 0.09ms<br/>99th percentile: 20.55ms ± 0.21ms<br/>Max time: 22.62ms ± 0.52ms | Ops: 102.32 ± 0.15 ops/s. Best: 102.50 ops/s<br/>Total: 9772.72ms ± 15.08ms. Best: 9756.19ms<br/>Min time: 1.44ms ± 0.11ms<br/>Mean time: 9.77ms ± 0.02ms<br/>Median time: 9.19ms ± 0.03ms<br/>95th percentile: 17.72ms ± 0.04ms<br/>99th percentile: 19.54ms ± 0.10ms<br/>Max time: 21.56ms ± 0.99ms |
| e2e_trip_mld | Ops: 56.08 ± 0.73 ops/s. Best: 57.06 ops/s<br/>Total: 17827.95ms ± 235.04ms. Best: 17524.72ms<br/>Min time: 1.93ms ± 0.24ms<br/>Mean time: 17.84ms ± 0.24ms<br/>Median time: 17.35ms ± 0.24ms<br/>95th percentile: 29.12ms ± 0.35ms<br/>99th percentile: 31.46ms ± 0.35ms<br/>Max time: 34.11ms ± 1.12ms | Ops: 60.55 ± 0.24 ops/s. Best: 60.83 ops/s<br/>Total: 16511.64ms ± 67.59ms. Best: 16439.61ms<br/>Min time: 1.55ms ± 0.23ms<br/>Mean time: 16.52ms ± 0.07ms<br/>Median time: 15.99ms ± 0.07ms<br/>95th percentile: 27.27ms ± 0.14ms<br/>99th percentile: 29.38ms ± 0.13ms<br/>Max time: 31.25ms ± 0.27ms |
| json-render | String: 5.61074ms<br/>Stringstream: 9.24026ms<br/>Vector: 6.60868ms | String: 5.62181ms<br/>Stringstream: 8.89788ms<br/>Vector: 6.51806ms |
| match_ch | Default radius: <br/>4.58919ms/req at 82 coordinate<br/>0.0559658ms/coordinate<br/>Radius 10m: <br/>16.0858ms/req at 82 coordinate<br/>0.196169ms/coordinate | Default radius: <br/>4.60838ms/req at 82 coordinate<br/>0.0561998ms/coordinate<br/>Radius 10m: <br/>16.1283ms/req at 82 coordinate<br/>0.196687ms/coordinate |
| match_mld | Default radius: <br/>3.01557ms/req at 82 coordinate<br/>0.0367752ms/coordinate<br/>Radius 10m: <br/>11.1467ms/req at 82 coordinate<br/>0.135935ms/coordinate | Default radius: <br/>3.04245ms/req at 82 coordinate<br/>0.037103ms/coordinate<br/>Radius 10m: <br/>11.5818ms/req at 82 coordinate<br/>0.141242ms/coordinate |
| osrm_contract | Time: 99.14s Peak RAM: 201.77MB | Time: 100.16s Peak RAM: 200.95MB |
| osrm_customize | Time: 1.29s Peak RAM: 117.55MB | Time: 1.30s Peak RAM: 117.69MB |
| osrm_extract | Time: 11.66s Peak RAM: 430.19MB | Time: 12.17s Peak RAM: 431.55MB |
| osrm_partition | Time: 2.08s Peak RAM: 144.07MB | Time: 2.13s Peak RAM: 145.80MB |
| packedvector | random write:<br/>std::vector 11311.5 ms<br/>util::packed_vector 73701.4 ms<br/>slowdown: 6.51561<br/>random read:<br/>std::vector 11050.1 ms<br/>util::packed_vector 30379.9 ms<br/>slowdown: 2.74929 | random write:<br/>std::vector 11262.9 ms<br/>util::packed_vector 74295.2 ms<br/>slowdown: 6.59643<br/>random read:<br/>std::vector 11122.7 ms<br/>util::packed_vector 30851.6 ms<br/>slowdown: 2.77375 |
| random_match_ch | 500 matches, default radius<br/>ops: 203.38 ± 0.84 ops/s. best: 204.20ops/s.<br/>total: 280.28 ± 1.16ms. best: 279.14ms.<br/>avg: 4.92 ± 0.02ms<br/>min: 0.13 ± 0.01ms<br/>max: 24.90 ± 0.04ms<br/>p99: 24.90 ± 0.04ms<br/><br/>500 matches, radius=10<br/>ops: 59.80 ± 0.09 ops/s. best: 59.95ops/s.<br/>total: 1070.26 ± 1.54ms. best: 1067.52ms.<br/>avg: 16.72 ± 0.02ms<br/>min: 0.15 ± 0.00ms<br/>max: 224.23 ± 1.12ms<br/>p99: 224.23 ± 1.12ms<br/><br/>500 matches, radius=20<br/>ops: 14.34 ± 0.02 ops/s. best: 14.36ops/s.<br/>total: 4531.84 ± 4.95ms. best: 4526.49ms.<br/>avg: 69.72 ± 0.08ms<br/>min: 0.30 ± 0.00ms<br/>max: 1142.35 ± 3.40ms<br/>p99: 1142.35 ± 3.40ms | 500 matches, default radius<br/>ops: 201.17 ± 0.73 ops/s. best: 202.13ops/s.<br/>total: 283.35 ± 1.03ms. best: 282.00ms.<br/>avg: 4.97 ± 0.02ms<br/>min: 0.14 ± 0.01ms<br/>max: 24.40 ± 0.03ms<br/>p99: 24.40 ± 0.03ms<br/><br/>500 matches, radius=10<br/>ops: 59.79 ± 0.25 ops/s. best: 60.29ops/s.<br/>total: 1070.47 ± 4.55ms. best: 1061.54ms.<br/>avg: 16.73 ± 0.07ms<br/>min: 0.15 ± 0.00ms<br/>max: 221.24 ± 1.20ms<br/>p99: 221.24 ± 1.20ms<br/><br/>500 matches, radius=20<br/>ops: 14.46 ± 0.05 ops/s. best: 14.56ops/s.<br/>total: 4494.10 ± 15.34ms. best: 4465.24ms.<br/>avg: 69.14 ± 0.24ms<br/>min: 0.30 ± 0.00ms<br/>max: 1131.12 ± 5.36ms<br/>p99: 1131.12 ± 5.36ms |
| random_match_mld | 500 matches, default radius<br/>ops: 294.25 ± 2.77 ops/s. best: 296.77ops/s.<br/>total: 193.74 ± 1.85ms. best: 192.07ms.<br/>avg: 3.40 ± 0.03ms<br/>min: 0.12 ± 0.01ms<br/>max: 19.57 ± 0.11ms<br/>p99: 19.57 ± 0.11ms<br/><br/>500 matches, radius=10<br/>ops: 104.01 ± 0.24 ops/s. best: 104.40ops/s.<br/>total: 615.35 ± 1.32ms. best: 613.00ms.<br/>avg: 9.61 ± 0.02ms<br/>min: 0.14 ± 0.00ms<br/>max: 114.19 ± 0.23ms<br/>p99: 114.19 ± 0.23ms<br/><br/>500 matches, radius=20<br/>ops: 21.04 ± 0.02 ops/s. best: 21.07ops/s.<br/>total: 3088.71 ± 3.00ms. best: 3085.53ms.<br/>avg: 47.52 ± 0.05ms<br/>min: 0.19 ± 0.00ms<br/>max: 601.29 ± 1.34ms<br/>p99: 601.29 ± 1.34ms | 500 matches, default radius<br/>ops: 303.25 ± 1.75 ops/s. best: 304.74ops/s.<br/>total: 187.98 ± 1.10ms. best: 187.04ms.<br/>avg: 3.30 ± 0.02ms<br/>min: 0.12 ± 0.01ms<br/>max: 18.91 ± 0.06ms<br/>p99: 18.91 ± 0.06ms<br/><br/>500 matches, radius=10<br/>ops: 106.93 ± 0.51 ops/s. best: 107.35ops/s.<br/>total: 598.54 ± 2.95ms. best: 596.18ms.<br/>avg: 9.35 ± 0.05ms<br/>min: 0.14 ± 0.00ms<br/>max: 110.71 ± 0.49ms<br/>p99: 110.71 ± 0.49ms<br/><br/>500 matches, radius=20<br/>ops: 21.76 ± 0.03 ops/s. best: 21.81ops/s.<br/>total: 2987.78 ± 3.60ms. best: 2980.02ms.<br/>avg: 45.97 ± 0.06ms<br/>min: 0.18 ± 0.01ms<br/>max: 582.86 ± 0.58ms<br/>p99: 582.86 ± 0.58ms |
| random_nearest_ch | 10000 nearest, number_of_results=1<br/>ops: 24647.59 ± 75.87 ops/s. best: 24710.24ops/s.<br/>total: 405.73 ± 1.28ms. best: 404.69ms.<br/>avg: 0.04 ± 0.00ms<br/>min: 0.01 ± 0.00ms<br/>max: 0.15 ± 0.03ms<br/>p99: 0.10 ± 0.00ms<br/><br/>10000 nearest, number_of_results=5<br/>ops: 18622.32 ± 40.91 ops/s. best: 18669.96ops/s.<br/>total: 536.99 ± 1.19ms. best: 535.62ms.<br/>avg: 0.05 ± 0.00ms<br/>min: 0.02 ± 0.00ms<br/>max: 0.15 ± 0.00ms<br/>p99: 0.11 ± 0.00ms<br/><br/>10000 nearest, number_of_results=10<br/>ops: 14793.61 ± 12.92 ops/s. best: 14820.43ops/s.<br/>total: 675.97 ± 0.59ms. best: 674.74ms.<br/>avg: 0.07 ± 0.00ms<br/>min: 0.03 ± 0.00ms<br/>max: 0.17 ± 0.00ms<br/>p99: 0.13 ± 0.00ms | 10000 nearest, number_of_results=1<br/>ops: 24335.03 ± 59.44 ops/s. best: 24429.17ops/s.<br/>total: 410.93 ± 1.07ms. best: 409.35ms.<br/>avg: 0.04 ± 0.00ms<br/>min: 0.01 ± 0.00ms<br/>max: 0.16 ± 0.03ms<br/>p99: 0.10 ± 0.00ms<br/><br/>10000 nearest, number_of_results=5<br/>ops: 18128.89 ± 48.45 ops/s. best: 18195.47ops/s.<br/>total: 551.61 ± 1.48ms. best: 549.59ms.<br/>avg: 0.06 ± 0.00ms<br/>min: 0.02 ± 0.00ms<br/>max: 0.23 ± 0.11ms<br/>p99: 0.11 ± 0.00ms<br/><br/>10000 nearest, number_of_results=10<br/>ops: 14427.20 ± 32.53 ops/s. best: 14482.95ops/s.<br/>total: 693.14 ± 1.53ms. best: 690.47ms.<br/>avg: 0.07 ± 0.00ms<br/>min: 0.03 ± 0.00ms<br/>max: 0.19 ± 0.01ms<br/>p99: 0.13 ± 0.00ms |
| random_nearest_mld | 10000 nearest, number_of_results=1<br/>ops: 24595.46 ± 51.44 ops/s. best: 24633.70ops/s.<br/>total: 406.58 ± 0.85ms. best: 405.95ms.<br/>avg: 0.04 ± 0.00ms<br/>min: 0.01 ± 0.00ms<br/>max: 0.15 ± 0.02ms<br/>p99: 0.10 ± 0.00ms<br/><br/>10000 nearest, number_of_results=5<br/>ops: 18560.48 ± 26.46 ops/s. best: 18619.55ops/s.<br/>total: 538.78 ± 0.77ms. best: 537.07ms.<br/>avg: 0.05 ± 0.00ms<br/>min: 0.02 ± 0.00ms<br/>max: 0.15 ± 0.00ms<br/>p99: 0.11 ± 0.00ms<br/><br/>10000 nearest, number_of_results=10<br/>ops: 14843.71 ± 12.08 ops/s. best: 14869.48ops/s.<br/>total: 673.69 ± 0.55ms. best: 672.52ms.<br/>avg: 0.07 ± 0.00ms<br/>min: 0.03 ± 0.00ms<br/>max: 0.17 ± 0.00ms<br/>p99: 0.13 ± 0.00ms | 10000 nearest, number_of_results=1<br/>ops: 24603.40 ± 48.80 ops/s. best: 24641.42ops/s.<br/>total: 406.45 ± 0.81ms. best: 405.82ms.<br/>avg: 0.04 ± 0.00ms<br/>min: 0.01 ± 0.00ms<br/>max: 0.16 ± 0.02ms<br/>p99: 0.10 ± 0.00ms<br/><br/>10000 nearest, number_of_results=5<br/>ops: 18512.69 ± 8.43 ops/s. best: 18532.01ops/s.<br/>total: 540.17 ± 0.24ms. best: 539.61ms.<br/>avg: 0.05 ± 0.00ms<br/>min: 0.02 ± 0.00ms<br/>max: 0.15 ± 0.00ms<br/>p99: 0.11 ± 0.00ms<br/><br/>10000 nearest, number_of_results=10<br/>ops: 14651.58 ± 7.81 ops/s. best: 14663.81ops/s.<br/>total: 682.52 ± 0.36ms. best: 681.95ms.<br/>avg: 0.07 ± 0.00ms<br/>min: 0.03 ± 0.00ms<br/>max: 0.18 ± 0.01ms<br/>p99: 0.13 ± 0.00ms |
| random_route_ch | 1000 routes, 3 coordinates, no alternatives, overview=full, steps=true<br/>ops: 536.28 ± 1.33 ops/s. best: 537.28ops/s.<br/>total: 1834.89 ± 4.56ms. best: 1831.45ms.<br/>avg: 1.86 ± 0.00ms<br/>min: 0.32 ± 0.00ms<br/>max: 3.07 ± 0.31ms<br/>p99: 2.63 ± 0.02ms<br/><br/>1000 routes, 2 coordinates, 3 alternatives, overview=full, steps=true<br/>ops: 650.31 ± 5.01 ops/s. best: 653.83ops/s.<br/>total: 1537.89 ± 12.02ms. best: 1529.44ms.<br/>avg: 1.54 ± 0.01ms<br/>min: 0.05 ± 0.00ms<br/>max: 3.90 ± 0.09ms<br/>p99: 3.19 ± 0.04ms<br/><br/>1000 routes, 3 coordinates, no alternatives, overview=false, steps=false<br/>ops: 1007.79 ± 14.25 ops/s. best: 1032.80ops/s.<br/>total: 976.63 ± 13.70ms. best: 952.75ms.<br/>avg: 0.99 ± 0.01ms<br/>min: 0.23 ± 0.00ms<br/>max: 1.57 ± 0.06ms<br/>p99: 1.36 ± 0.02ms<br/><br/>1000 routes, 2 coordinates, 3 alternatives, overview=false, steps=false<br/>ops: 1233.32 ± 5.00 ops/s. best: 1237.98ops/s.<br/>total: 810.84 ± 3.31ms. best: 807.76ms.<br/>avg: 0.81 ± 0.00ms<br/>min: 0.05 ± 0.00ms<br/>max: 2.08 ± 0.03ms<br/>p99: 1.74 ± 0.03ms | 1000 routes, 3 coordinates, no alternatives, overview=full, steps=true<br/>ops: 478.57 ± 5.53 ops/s. best: 489.51ops/s.<br/>total: 2056.49 ± 23.83ms. best: 2010.19ms.<br/>avg: 2.09 ± 0.02ms<br/>min: 0.33 ± 0.01ms<br/>max: 3.79 ± 0.23ms<br/>p99: 3.09 ± 0.03ms<br/><br/>1000 routes, 2 coordinates, 3 alternatives, overview=full, steps=true<br/>ops: 579.28 ± 14.17 ops/s. best: 593.82ops/s.<br/>total: 1727.82 ± 43.65ms. best: 1684.02ms.<br/>avg: 1.73 ± 0.04ms<br/>min: 0.06 ± 0.00ms<br/>max: 4.58 ± 0.17ms<br/>p99: 3.81 ± 0.13ms<br/><br/>1000 routes, 3 coordinates, no alternatives, overview=false, steps=false<br/>ops: 977.51 ± 15.15 ops/s. best: 992.54ops/s.<br/>total: 1006.95 ± 15.75ms. best: 991.40ms.<br/>avg: 1.02 ± 0.02ms<br/>min: 0.23 ± 0.00ms<br/>max: 1.67 ± 0.04ms<br/>p99: 1.43 ± 0.04ms<br/><br/>1000 routes, 2 coordinates, 3 alternatives, overview=false, steps=false<br/>ops: 969.28 ± 43.22 ops/s. best: 1023.54ops/s.<br/>total: 1034.57 ± 46.87ms. best: 977.00ms.<br/>avg: 1.03 ± 0.05ms<br/>min: 0.04 ± 0.00ms<br/>max: 2.69 ± 0.13ms<br/>p99: 2.38 ± 0.12ms |
| random_route_mld | 1000 routes, 3 coordinates, no alternatives, overview=full, steps=true<br/>ops: 261.21 ± 0.47 ops/s. best: 261.72ops/s.<br/>total: 3767.08 ± 6.83ms. best: 3759.69ms.<br/>avg: 3.83 ± 0.01ms<br/>min: 0.31 ± 0.00ms<br/>max: 8.45 ± 0.04ms<br/>p99: 6.42 ± 0.06ms<br/><br/>1000 routes, 2 coordinates, 3 alternatives, overview=full, steps=true<br/>ops: 252.75 ± 0.16 ops/s. best: 253.03ops/s.<br/>total: 3956.52 ± 2.55ms. best: 3952.13ms.<br/>avg: 3.96 ± 0.00ms<br/>min: 0.05 ± 0.00ms<br/>max: 9.25 ± 0.21ms<br/>p99: 8.18 ± 0.02ms<br/><br/>1000 routes, 3 coordinates, no alternatives, overview=false, steps=false<br/>ops: 343.10 ± 0.31 ops/s. best: 343.48ops/s.<br/>total: 2868.01 ± 2.62ms. best: 2864.82ms.<br/>avg: 2.91 ± 0.00ms<br/>min: 0.26 ± 0.00ms<br/>max: 7.11 ± 0.02ms<br/>p99: 5.01 ± 0.02ms<br/><br/>1000 routes, 2 coordinates, 3 alternatives, overview=false, steps=false<br/>ops: 305.63 ± 5.84 ops/s. best: 313.58ops/s.<br/>total: 3273.37 ± 62.35ms. best: 3188.98ms.<br/>avg: 3.27 ± 0.06ms<br/>min: 0.04 ± 0.00ms<br/>max: 7.13 ± 0.29ms<br/>p99: 6.53 ± 0.11ms | 1000 routes, 3 coordinates, no alternatives, overview=full, steps=true<br/>ops: 260.41 ± 1.19 ops/s. best: 261.60ops/s.<br/>total: 3778.75 ± 17.35ms. best: 3761.48ms.<br/>avg: 3.84 ± 0.02ms<br/>min: 0.31 ± 0.00ms<br/>max: 8.47 ± 0.02ms<br/>p99: 6.41 ± 0.04ms<br/><br/>1000 routes, 2 coordinates, 3 alternatives, overview=full, steps=true<br/>ops: 246.78 ± 2.73 ops/s. best: 252.88ops/s.<br/>total: 4052.95 ± 44.25ms. best: 3954.41ms.<br/>avg: 4.05 ± 0.04ms<br/>min: 0.05 ± 0.00ms<br/>max: 10.84 ± 1.44ms<br/>p99: 8.45 ± 0.20ms<br/><br/>1000 routes, 3 coordinates, no alternatives, overview=false, steps=false<br/>ops: 334.59 ± 7.22 ops/s. best: 341.47ops/s.<br/>total: 2942.99 ± 65.31ms. best: 2881.66ms.<br/>avg: 2.99 ± 0.07ms<br/>min: 0.27 ± 0.00ms<br/>max: 7.11 ± 0.01ms<br/>p99: 5.19 ± 0.11ms<br/><br/>1000 routes, 2 coordinates, 3 alternatives, overview=false, steps=false<br/>ops: 310.95 ± 1.99 ops/s. best: 313.32ops/s.<br/>total: 3216.17 ± 20.70ms. best: 3191.65ms.<br/>avg: 3.22 ± 0.02ms<br/>min: 0.04 ± 0.00ms<br/>max: 6.82 ± 0.08ms<br/>p99: 6.44 ± 0.03ms |
| random_table_ch | 250 tables, 3 coordinates<br/>ops: 1447.34 ± 11.61 ops/s. best: 1456.46ops/s.<br/>total: 172.75 ± 1.41ms. best: 171.65ms.<br/>avg: 0.69 ± 0.01ms<br/>min: 0.46 ± 0.00ms<br/>max: 1.05 ± 0.23ms<br/>p99: 0.86 ± 0.02ms<br/><br/>250 tables, 25 coordinates<br/>ops: 169.74 ± 0.03 ops/s. best: 169.79ops/s.<br/>total: 1472.80 ± 0.29ms. best: 1472.41ms.<br/>avg: 5.89 ± 0.00ms<br/>min: 5.21 ± 0.02ms<br/>max: 6.48 ± 0.07ms<br/>p99: 6.38 ± 0.01ms<br/><br/>250 tables, 50 coordinates<br/>ops: 84.04 ± 0.13 ops/s. best: 84.18ops/s.<br/>total: 2974.79 ± 4.71ms. best: 2969.66ms.<br/>avg: 11.90 ± 0.02ms<br/>min: 11.01 ± 0.02ms<br/>max: 12.93 ± 0.05ms<br/>p99: 12.74 ± 0.04ms | 250 tables, 3 coordinates<br/>ops: 1359.30 ± 14.56 ops/s. best: 1376.65ops/s.<br/>total: 183.95 ± 2.03ms. best: 181.60ms.<br/>avg: 0.74 ± 0.01ms<br/>min: 0.50 ± 0.01ms<br/>max: 1.15 ± 0.31ms<br/>p99: 0.93 ± 0.05ms<br/><br/>250 tables, 25 coordinates<br/>ops: 160.72 ± 1.09 ops/s. best: 162.37ops/s.<br/>total: 1555.62 ± 10.53ms. best: 1539.73ms.<br/>avg: 6.22 ± 0.04ms<br/>min: 5.62 ± 0.07ms<br/>max: 6.87 ± 0.05ms<br/>p99: 6.75 ± 0.05ms<br/><br/>250 tables, 50 coordinates<br/>ops: 79.30 ± 0.53 ops/s. best: 80.03ops/s.<br/>total: 3152.73 ± 22.03ms. best: 3123.88ms.<br/>avg: 12.61 ± 0.09ms<br/>min: 11.75 ± 0.08ms<br/>max: 13.72 ± 0.13ms<br/>p99: 13.48 ± 0.13ms |
…